Move and rename mckctl under ci/ as mckci
