+++ ci/cmd/mckci/main.go
@@ -1,9 +1,9 @@
-// mckctl is the MongoDB Controllers for Kubernetes (MCK) developer tooling
-// entry point. The current scope is release automation under `mckctl release`;
+// mckci is the MongoDB Controllers for Kubernetes (MCK) developer tooling
+// entry point. The current scope is release automation under `mckci release`;
 // other domains (CI, Evergreen, etc.) may register top-level subcommands later.
 //
 // This is a thin entry point: it sets up signal handling and delegates to
-// internal/ci/cli. All runnable logic lives under internal/ci so it can be
+// ci/internal/cli. All runnable logic lives under ci/internal so it can be
 // imported from tests and so it never ends up linked into the operator binary.
 package main
 
@@ -13,7 +13,7 @@ import (
 	"os/signal"
 	"syscall"
 
-	"github.com/mongodb/mongodb-kubernetes/internal/ci/cli"
+	"github.com/mongodb/mongodb-kubernetes/ci/internal/cli"
 )
 
 func main() {
