+++ scripts/mckci
@@ -1,0 +1,10 @@
+#!/usr/bin/env bash
+# scripts/mckci — build mckci from source if stale, then exec it.
+#
+# Lets users invoke `scripts/mckci release pr ...` without remembering to
+# rebuild after pulling. Go's build cache makes incremental rebuilds near
+# instant; only the first build is slow.
+set -euo pipefail
+cd "$(git rev-parse --show-toplevel)"
+go build -o bin/mckci ./ci/cmd/mckci
+exec ./bin/mckci "$@"
